Automatically register HTTP HEAD webscripts
The HTTP HEAD method is supposed to return the same headers as GET on the same path,
as per RFC2616, section 9.4. (https://tools.ietf.org/html/rfc2616#section-9.4)

Since it is not possible to manually create a dynamic extensions
webscript that handles HEAD requests, we can create them automatically
for webscripts that handle GET requests.